move AuthCreds to internaloption
