+++ transport/dial.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"net/http"
 
-	"cloud.google.com/go/auth"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/grpc"
 
@@ -46,31 +45,4 @@ func Creds(ctx context.Context, opts ...option.ClientOption) (*google.Credential
 		opt.Apply(&ds)
 	}
 	return internal.Creds(ctx, &ds)
-}
-
-// AuthCreds returns [cloud.google.com/go/auth.Credentials] using the following
-// options provided via [option.ClientOption], including legacy oauth2/google
-// options, in this order:
-//
-// * [option.WithAuthCredentials]
-// * [option/internaloption.WithCredentials] (internal use only)
-// * [option.WithCredentials]
-// * [option.WithTokenSource]
-//
-// If there are no applicable credentials options, then it passes the
-// following options to [cloud.google.com/go/auth/credentials.DetectDefault] and
-// returns the result:
-//
-// * [option.WithAudiences]
-// * [option.WithCredentialsFile]
-// * [option.WithCredentialsJSON]
-// * [option.WithScopes]
-// * [option/internaloption.WithDefaultScopes] (internal use only)
-// * [option/internaloption.EnableJwtWithScope] (internal use only)
-func AuthCreds(ctx context.Context, opts ...option.ClientOption) (*auth.Credentials, error) {
-	var ds internal.DialSettings
-	for _, opt := range opts {
-		opt.Apply(&ds)
-	}
-	return internal.AuthCreds(ctx, &ds)
 }
